backend/requirements.txt
Django==3.2.18
django-cors-headers==3.13.0
djangorestframework==3.14.0
multipart==0.2.4
nltk==3.8.1
openai==0.27.8
pandas==1.3.5
Pillow==9.4.0
PyJWT==2.4.0
regex==2022.10.31
requests==2.31.0
djangorestframework-simplejwt==5.2.2
fpdf
django-grip
pymongo[srv]==3.12.1
django-environ==0.10.0
azure-storage-blob==12.17.0
channels[daphne]==3.0.4
channels-redis==4.1.0
PyPDF2==3.0.1
pdfservices-sdk==2.3.0
matplotlib==3.7.2
scikit-learn==1.3.0
plotly==5.15.0
tiktoken==0.4.0
PyMuPDF==1.21.0
reportlab==4.0.4
loguru==0.7.0
openpyxl==3.1.2
python-docx==0.8.11
# django-eventstream==4.5.1
xlrd==2.0.1
textract==1.6.5
python-environ==0.4.54
redis==4.6.0
celery
pdf2image
docx2pdf
pytesseract